[6123] Improve PaletteToolList rendering
Bug: https://github.com/eclipse-sirius/sirius-web/issues/6123
