+++ packages/palette/frontend/sirius-components-palette/src/tool-section/PaletteToolSection.types.ts
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2024, 2025 Obeo.
+ * Copyright (c) 2026 Obeo.
  * This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v2.0
  * which accompanies this distribution, and is available at
@@ -11,20 +11,21 @@
  *     Obeo - initial API and implementation
  *******************************************************************************/
 
-import { GQLPalette, GQLTool, GQLToolSection } from '../Palette.types';
+import { GQLPalette, GQLTool } from '../Palette.types';
 import { PaletteExtensionSectionProps } from '../PaletteExtensionSection.types';
 
-export interface PaletteToolListProps {
+export interface PaletteToolSectionProps {
   onToolClick: (tool: GQLTool) => void;
   palette: GQLPalette;
   onBackToMainList: () => void;
   onClose: () => void;
   representationElementIds: string[];
   lastToolInvoked: GQLTool | null;
-  children: React.ReactElement<PaletteExtensionSectionProps>[];
+  extensionSections: React.ReactElement<PaletteExtensionSectionProps>[];
 }
 
-export interface PaletteToolListStateValue {
-  toolSection: GQLToolSection | null;
-  extensionSection: string | null;
+export interface PaletteToolSectionStateValue {
+  currentSectionId: string | undefined;
+  history: (string | undefined)[];
+  direction: 'left' | 'right';
 }
