+++ packages/palette/frontend/sirius-components-palette/src/tool-section/ToolSectionHeader.types.ts
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2024, 2025 Obeo.
+ * Copyright (c) 2026 Obeo.
  * This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v2.0
  * which accompanies this distribution, and is available at
@@ -10,11 +10,8 @@
  * Contributors:
  *     Obeo - initial API and implementation
  *******************************************************************************/
-import { GQLTool, GQLToolSection } from '../Palette.types';
 
-export interface PaletteToolSectionListProps {
-  onToolClick: (tool: GQLTool) => void;
-  onBackToMainList: () => void;
-  toolSection: GQLToolSection;
-  representationElementIds: string[];
+export interface ToolSectionHeaderProps {
+  title: string;
+  navigateBack: () => void;
 }
